backend/requirements.txt
mysql-connector-python
deepface
flask
opencv-python
numpy
